fix(vaults): improvements after the review
- transferVaultOwnership instead of setVaultOwner
- comments and naming improvements
- fix a bug when dashboard initialization is failing with ReportStale

+++ contracts/0.8.9/oracle/AccountingOracle.sol
@@ -19,7 +19,7 @@ interface IReportReceiver {
 
 interface ILazyOracle {
     function updateReportData(
-        uint64 _timestamp,
+        uint256 _timestamp,
         bytes32 _vaultsDataTreeRoot,
         string memory _vaultsDataReportCid
     ) external;
@@ -508,7 +508,7 @@ contract AccountingOracle is BaseOracle {
         );
 
         ILazyOracle(LOCATOR.lazyOracle()).updateReportData(
-            uint64(GENESIS_TIME + data.refSlot * SECONDS_PER_SLOT),
+            GENESIS_TIME + data.refSlot * SECONDS_PER_SLOT,
             data.vaultsDataTreeRoot,
             data.vaultsDataTreeCid
         );
